providers/oracle: move oracle-oci to oci mod
this makes it more clear organizationally that this metadata fetcher is
for the oracle-oci metadata service, and brings this module in line with
other providers with mutliple metadata sources, like cloudstack and
openstack.
